bot/requirements.txt.txt
python-telegram-bot==20.7
sqlite3
requests
python-dotenv